Service control feature implemented

+++ NativeDllTestApp/NativeAPI.cs
@@ -11,10 +11,10 @@ namespace NativeDllTestApp
         private const string DLL_LOCATION = @"x64\svcman.dll";
 
         [DllImport(DLL_LOCATION, CallingConvention = CallingConvention.Cdecl)]
-        public static extern ulong ServiceCount();
+        public static extern uint ServiceCount();
        
         [DllImport(DLL_LOCATION, CallingConvention = CallingConvention.Cdecl)]
-        public static extern void ServiceList([In, Out] ServiceProcess[] services, ulong count);
+        public static extern void ServiceList([In, Out] ServiceProcess[] services, uint count);
 
         [DllImport(DLL_LOCATION, CharSet = CharSet.Unicode, CallingConvention = CallingConvention.StdCall)]
         public static extern int StartWinService([MarshalAs(UnmanagedType.BStr)] string serviceName, ref ServiceControlState svcState);
